SAE-Aerothon-2025/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
opencv-python
numpy
requests
pydantic
Pillow
google-generativeai
shapely
matplotlib
pymavlink
pyserial
sqlalchemy